+++ src/renderer/modals/ExchangeDeviceConfirm/index.js
@@ -168,9 +168,26 @@ const BuyCrypto = () => {
       centered
       render={({ data, onClose }) => (
         <ModalBody
-          onClose={onClose}
+          onClose={() => {
+            if (data.onCancel) {
+              data.onCancel();
+            }
+            onClose();
+          }}
           title="Connect your device"
-          render={() => (data ? <Root data={data} onClose={onClose} /> : null)}
+          render={() =>
+            data ? (
+              <Root
+                data={data}
+                onClose={() => {
+                  if (data.onCancel) {
+                    data.onCancel();
+                  }
+                  onClose();
+                }}
+              />
+            ) : null
+          }
         />
       )}
     />
