--- src/renderer/modals/SellDeviceConfirm/Body.js
@@ -1,261 +1,0 @@
-// @flow
-
-import React, { useCallback, useEffect, useState } from "react";
-import { BigNumber } from "bignumber.js";
-import { connect } from "react-redux";
-import { compose } from "redux";
-import type { TFunction } from "react-i18next";
-import { createStructuredSelector } from "reselect";
-import { Trans, withTranslation } from "react-i18next";
-import { UserRefusedOnDevice } from "@ledgerhq/errors";
-import { addPendingOperation, getMainAccount } from "@ledgerhq/live-common/lib/account";
-import useBridgeTransaction from "@ledgerhq/live-common/lib/bridge/useBridgeTransaction";
-import type { Account, AccountLike, Operation } from "@ledgerhq/live-common/lib/types";
-import logger from "~/logger";
-import Stepper from "~/renderer/components/Stepper";
-import { SyncSkipUnderPriority } from "@ledgerhq/live-common/lib/bridge/react";
-import { closeModal, openModal } from "~/renderer/actions/modals";
-import { accountsSelector } from "~/renderer/reducers/accounts";
-import { updateAccountWithUpdater } from "~/renderer/actions/accounts";
-import { getCurrentDevice } from "~/renderer/reducers/devices";
-import Track from "~/renderer/analytics/Track";
-import type { Device } from "@ledgerhq/live-common/lib/hw/actions/types";
-import StepConnectDevice from "./steps/StepConnectDevice";
-import StepSummary, { StepSummaryFooter } from "./steps/StepSummary";
-import StepConfirmation, { StepConfirmationFooter } from "./steps/StepConfirmation";
-import StepWarning, { StepWarningFooter } from "./steps/StepWarning";
-import type { St, StepId } from "./types";
-import { getAccountBridge } from "@ledgerhq/live-common/lib/bridge";
-
-type OwnProps = {|
-  stepId: StepId,
-  onChangeStepId: StepId => void,
-  onClose: () => void,
-  params: {
-    account: ?AccountLike,
-    parentAccount: ?Account,
-    startWithWarning?: boolean,
-    recipient?: string,
-    amount?: BigNumber,
-    onChangeTransactionId: string => void,
-  },
-|};
-
-type StateProps = {|
-  t: TFunction,
-  device: ?Device,
-  accounts: Account[],
-  closeModal: string => void,
-  openModal: (string, any) => void,
-  updateAccountWithUpdater: (string, (Account) => Account) => void,
-|};
-
-type Props = {|
-  ...OwnProps,
-  ...StateProps,
-|};
-
-const createSteps = (): St[] => [
-  {
-    id: "warning",
-    excludeFromBreadcrumb: true,
-    component: StepWarning,
-    footer: StepWarningFooter,
-  },
-  {
-    id: "summary",
-    label: <Trans i18nKey="send.steps.summary.title" />,
-    component: StepSummary,
-    footer: StepSummaryFooter,
-  },
-  {
-    id: "device",
-    label: <Trans i18nKey="send.steps.device.title" />,
-    component: StepConnectDevice,
-    onBack: ({ transitionTo }) => transitionTo("summary"),
-  },
-  {
-    id: "confirmation",
-    label: <Trans i18nKey="send.steps.confirmation.title" />,
-    excludeFromBreadcrumb: true,
-    component: StepConfirmation,
-    footer: StepConfirmationFooter,
-  },
-];
-
-const mapStateToProps = createStructuredSelector({
-  device: getCurrentDevice,
-  accounts: accountsSelector,
-});
-
-const mapDispatchToProps = {
-  closeModal,
-  openModal,
-  updateAccountWithUpdater,
-};
-
-const Body = ({
-  t,
-  device,
-  openModal,
-  closeModal,
-  onChangeStepId,
-  onClose,
-  stepId,
-  params,
-  accounts,
-  updateAccountWithUpdater,
-}: Props) => {
-  const openedFromAccount = !!params.account;
-  const [steps] = useState(createSteps);
-
-  // initial values might coming from deeplink
-  const [maybeAmount, setMaybeAmount] = useState(() => params.amount || null);
-  const [maybeRecipient, setMaybeRecipient] = useState(() => params.recipient || null);
-
-  const onResetMaybeAmount = useCallback(() => {
-    setMaybeAmount(null);
-  }, [setMaybeAmount]);
-
-  const onResetMaybeRecipient = useCallback(() => {
-    setMaybeRecipient(null);
-  }, [setMaybeRecipient]);
-
-  const {
-    transaction,
-    setTransaction,
-    updateTransaction,
-    account,
-    parentAccount,
-    setAccount,
-    status,
-    bridgeError,
-    bridgePending,
-  } = useBridgeTransaction(() => {
-    const parentAccount = params && params.parentAccount;
-    const account = (params && params.account) || accounts[0];
-    const mainAccount = getMainAccount(account, parentAccount);
-
-    const bridge = getAccountBridge(account, parentAccount);
-
-    const t = bridge.createTransaction(mainAccount);
-
-    const transaction = bridge.updateTransaction(t, {
-      amount: params.amount,
-      recipient: params.recipient,
-    });
-
-    return { account, parentAccount, transaction };
-  });
-
-  // make sure step id is in sync
-  useEffect(() => {
-    const stepId = params && params.startWithWarning ? "warning" : null;
-    if (stepId) onChangeStepId(stepId);
-  }, [onChangeStepId, params]);
-
-  const [optimisticOperation, setOptimisticOperation] = useState(null);
-  const [transactionError, setTransactionError] = useState(null);
-  const [signed, setSigned] = useState(false);
-
-  const handleCloseModal = useCallback(() => {
-    closeModal("MODAL_SELL_CRYPTO_DEVICE");
-  }, [closeModal]);
-
-  const handleChangeAccount = useCallback(
-    (nextAccount: AccountLike, nextParentAccount: ?Account) => {
-      if (account !== nextAccount) {
-        setAccount(nextAccount, nextParentAccount);
-      }
-    },
-    [account, setAccount],
-  );
-
-  const handleRetry = useCallback(() => {
-    setTransactionError(null);
-    setOptimisticOperation(null);
-    setSigned(false);
-  }, []);
-
-  const handleTransactionError = useCallback((error: Error) => {
-    if (!(error instanceof UserRefusedOnDevice)) {
-      logger.critical(error);
-    }
-    setTransactionError(error);
-  }, []);
-
-  const handleOperationBroadcasted = useCallback(
-    (optimisticOperation: Operation) => {
-      if (!account) return;
-      const mainAccount = getMainAccount(account, parentAccount);
-      updateAccountWithUpdater(mainAccount.id, account =>
-        addPendingOperation(account, optimisticOperation),
-      );
-      setOptimisticOperation(optimisticOperation);
-      setTransactionError(null);
-    },
-    [account, parentAccount, updateAccountWithUpdater],
-  );
-
-  const handleStepChange = useCallback(e => onChangeStepId(e.id), [onChangeStepId]);
-
-  const errorSteps = [];
-
-  if (transactionError) {
-    errorSteps.push(3);
-  } else if (bridgeError) {
-    errorSteps.push(0);
-  }
-
-  const error = transactionError || bridgeError;
-
-  const stepperProps = {
-    title: stepId === "warning" ? t("common.information") : t("send.title"),
-    stepId,
-    steps,
-    errorSteps,
-    device,
-    openedFromAccount,
-    account,
-    parentAccount,
-    transaction,
-    signed,
-    hideBreadcrumb: (!!error && ["recipient", "amount"].includes(stepId)) || stepId === "warning",
-    error,
-    status,
-    bridgePending,
-    optimisticOperation,
-    openModal,
-    onClose,
-    setSigned,
-    closeModal: handleCloseModal,
-    onChangeAccount: handleChangeAccount,
-    onChangeTransaction: setTransaction,
-    onRetry: handleRetry,
-    onStepChange: handleStepChange,
-    onOperationBroadcasted: handleOperationBroadcasted,
-    onTransactionError: handleTransactionError,
-    setTransactionId: params.onChangeTransactionId,
-    maybeAmount,
-    onResetMaybeAmount,
-    maybeRecipient,
-    onResetMaybeRecipient,
-    updateTransaction,
-  };
-
-  if (!status) return null;
-
-  return (
-    <Stepper {...stepperProps}>
-      <SyncSkipUnderPriority priority={100} />
-      <Track onUnmount event="CloseModalSend" />
-    </Stepper>
-  );
-};
-
-const m: React$ComponentType<OwnProps> = compose(
-  connect(mapStateToProps, mapDispatchToProps),
-  withTranslation(),
-)(Body);
-
-export default m;
