--- src/renderer/modals/SellDeviceConfirm/steps/StepConnectDevice.js
@@ -1,34 +1,0 @@
-// @flow
-import React from "react";
-import TrackPage from "~/renderer/analytics/TrackPage";
-import GenericStepConnectDevice from "./GenericStepConnectDevice";
-import type { StepProps } from "../types";
-
-export default function StepConnectDevice({
-  account,
-  parentAccount,
-  transaction,
-  status,
-  transitionTo,
-  onOperationBroadcasted,
-  onTransactionError,
-  setSigned,
-  setTransactionId,
-}: StepProps) {
-  return (
-    <>
-      <TrackPage category="Send Flow" name="Step ConnectDevice" />
-      <GenericStepConnectDevice
-        account={account}
-        parentAccount={parentAccount}
-        transaction={transaction}
-        status={status}
-        transitionTo={transitionTo}
-        onOperationBroadcasted={onOperationBroadcasted}
-        onTransactionError={onTransactionError}
-        setSigned={setSigned}
-        setTransactionId={setTransactionId}
-      />
-    </>
-  );
-}
